live-trading/requirements.txt
TikTokLive>=5.0.0
flask>=3.0.0
flask-cors>=4.0.0
pyautogui>=0.9.54
pygetwindow>=0.0.9
requests>=2.31.0
python-dotenv>=1.0.0
